requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.0.2
django-heroku==0.3.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
lxml==4.4.1
nflgame-redux==2.0.1a1
numpy==1.17.1
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
sql==0.4.0
SQLAlchemy==1.3.8
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
wcwidth==0.1.7
whitenoise==4.1.3
